fix: remove circular deps and ci check
