+++ package.json
@@ -8,12 +8,13 @@
     "url": "https://github.com/angular/material2.git"
   },
   "scripts": {
-    "ci:forbidden-identifiers": "node ./scripts/ci/forbidden-identifiers.js",    
+    "ci:forbidden-identifiers": "node ./scripts/ci/forbidden-identifiers.js",
     "build": "ng build",
     "demo-app": "ng serve",
     "test": "karma start test/karma.conf.js",
     "tslint": "tslint -c tslint.json 'src/**/*.ts'",
     "stylelint": "stylelint 'src/**/*.scss' --config stylelint-config.json --syntax scss",
+    "check-circular-deps": "madge --circular ./dist",
     "typings": "typings install --ambient",
     "postinstall": "npm run typings",
     "e2e": "protractor",
@@ -60,13 +61,14 @@
     "karma-firefox-launcher": "^0.1.7",
     "karma-jasmine": "^0.3.8",
     "karma-sauce-launcher": "^0.2.14",
+    "madge": "^0.5.3",
     "node-sass": "^3.4.2",
     "protractor": "^3.3.0",
     "protractor-accessibility-plugin": "0.1.1",
     "sass": "^0.5.0",
     "strip-ansi": "^3.0.0",
-    "symlink-or-copy": "^1.0.1",
     "stylelint": "^6.5.1",
+    "symlink-or-copy": "^1.0.1",
     "ts-node": "^0.7.3",
     "tslint": "^3.5.0",
     "typescript": "^1.9.0-dev",
