+++ scripts/ci/build-and-test.sh
@@ -18,7 +18,9 @@ wait_for_tunnel
 if is_lint; then
   npm run tslint  
   npm run ci:forbidden-identifiers
-  npm run stylelint  
+  npm run stylelint
+elif is_circular_deps_check; then
+  npm run check-circular-deps
 elif is_e2e; then
   MD_APP=e2e ng serve &
   sleep 20
