+++ src/core/a11y/live-announcer.spec.ts
