+++ src/demo-app/main.scss
@@ -3,4 +3,4 @@
 // that are consumed across multiple components (and thus shouldn't be scoped).
 
 @import "core/overlay/overlay";
-@import "core/live-announcer/live-announcer";
+@import "core/a11y/live-announcer";
