+++ src/e2e-app/main.scss
@@ -1,2 +1,2 @@
 @import "core/overlay/overlay";
-@import "core/live-announcer/live-announcer";
+@import "core/a11y/live-announcer";
