+++ src/e2e-app/main.ts
@@ -6,7 +6,7 @@ import {provide} from '@angular/core';
 import {Renderer} from '@angular/core';
 
 import {OVERLAY_CONTAINER_TOKEN} from '@angular2-material/core/overlay/overlay';
-import {MdLiveAnnouncer} from '@angular2-material/core/live-announcer/live-announcer';
+import {MdLiveAnnouncer} from '@angular2-material/core/a11y/live-announcer';
 import {createOverlayContainer} from '@angular2-material/core/overlay/overlay-container';
 import {MdGestureConfig} from '@angular2-material/core/gestures/MdGestureConfig';
 import {MdIconRegistry} from '@angular2-material/icon/icon-registry';
